The new Ping sensor is *much* thicker.
I measured again, and it needs a solid 4mm of vertical space to avoid hitting the battery pack. This change shouldn't affect the bottom half at all, but the top is now about 2.5mm taller. I switched to M3x25 screws, and changed the shape of the top face slightly to give the screw holes a good place to sit.